Added codeql and beta publishing actions, updated dependencies
