+++ .github/workflows/ci.yml
@@ -1,12 +1,12 @@
 name: CI
 
 on:
-  push:
-    branches:
-    - "main"
   pull_request:
     branches:
-    - "main"
+      - 'main'
+
+env:
+  CI: true
 
 jobs:
   pullRequestBuild:
@@ -14,7 +14,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        node-version: [16.x]
+        node-version: [16.x, 18.x]
     steps:
       - uses: actions/checkout@v3
       - name: Use Node.js ${{ matrix.node-version }}
@@ -23,33 +23,24 @@ jobs:
           node-version: ${{ matrix.node-version }}
           cache: 'npm'
       - name: Install dependencies
-        run: npm ci
-        env:
-          NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
-          CI: true
+        run: npm ci --ignore-scripts
+      - name: Audit Signatures
+        run: npm audit signatures
       - name: Compile
         run: npm run ci:compile
-        env:
-          CI: true
       - name: Check Code Style
         run: npm run ci:style
-        env:
-          CI: true
       - name: Perform Linting
         run: npm run ci:lint
-        env:
-          CI: true
       - name: Run Tests
         run: npm run ci:test
-        env:
-          CI: true
 
   branchBuild:
     runs-on: ubuntu-latest
     name: Branch Build
     strategy:
       matrix:
-        node-version: [16.x]
+        node-version: [16.x, 18.x]
     steps:
       - uses: actions/checkout@v3
         with:
@@ -60,23 +51,14 @@ jobs:
           node-version: ${{ matrix.node-version }}
           cache: 'npm'
       - name: Install dependencies
-        run: npm ci
-        env:
-          NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
-          CI: true
+        run: npm ci --ignore-scripts
+      - name: Audit Signatures
+        run: npm audit signatures
       - name: Compile
         run: npm run ci:compile
-        env:
-          CI: true
       - name: Check Code Style
         run: npm run ci:style
-        env:
-          CI: true
       - name: Perform Linting
         run: npm run ci:lint
-        env:
-          CI: true
       - name: Run Tests
         run: npm run ci:test
-        env:
-          CI: true
