+++ SECURITY.md
@@ -1,0 +1,14 @@
+# Security Policy
+
+## Supported Versions
+
+These versions of `@checkdigit/retry` are currently being supported with security updates.
+
+| Version | Supported          |
+| ------- | ------------------ |
+| \>= 2.0 | :white_check_mark: |
+| \< 2.0  | :x:                |
+
+## Reporting a Vulnerability
+
+Please create an issue at https://github.com/checkdigit/retry/issues
