+++ package.json
@@ -19,44 +19,35 @@
   },
   "homepage": "https://github.com/checkdigit/timeout#readme",
   "devDependencies": {
-    "@checkdigit/eslint-config": "^7.0.0",
-    "@checkdigit/prettier-config": "^2.1.0",
-    "@checkdigit/typescript-config": "3.0.1",
-    "@types/jest": "^28.1.3",
-    "jest": "^28.1.1",
-    "rimraf": "^3.0.2",
-    "ts-jest": "^28.0.5",
-    "ts-node": "^10.8.1"
-  },
-  "jest": {
-    "preset": "ts-jest",
-    "collectCoverageFrom": [
-      "<rootDir>/src/**",
-      "!<rootDir>/src/**/*.spec.ts",
-      "!<rootDir>/src/**/*.test.ts"
-    ],
-    "globals": {
-      "ts-jest": {
-        "isolatedModules": true,
-        "diagnostics": false
-      }
-    },
-    "testMatch": [
-      "<rootDir>/src/**/*.spec.ts"
-    ]
+    "@checkdigit/eslint-config": "^7.8.0",
+    "@checkdigit/jest-config": "^2.2.1",
+    "@checkdigit/prettier-config": "^3.4.0",
+    "@checkdigit/typescript-config": "3.3.0"
   },
   "scripts": {
     "preversion": "npm test",
     "postversion": "git push && git push --tags",
-    "lint": "eslint -f unix 'src/**/*.ts'",
-    "lint:fix": "eslint -f unix 'src/**/*.ts' --fix",
-    "prettier": "prettier --list-different 'src/**/*.ts'",
-    "prettier:fix": "prettier --write 'src/**/*.ts'",
+    "lint:fix": "eslint -f unix --ext .ts,.mts src --fix",
+    "lint": "eslint -f unix --ext .ts,.mts src",
+    "prettier": "prettier --ignore-path .gitignore --list-different .",
+    "prettier:fix": "prettier --ignore-path .gitignore --write .",
     "test": "npm run ci:compile && npm run ci:test && npm run ci:lint && npm run ci:style",
     "ci:compile": "tsc --noEmit",
     "ci:test": "jest --coverage=false",
     "ci:coverage": "jest --coverage=true",
     "ci:lint": "npm run lint",
     "ci:style": "npm run prettier"
-  }
+  },
+  "jest": {
+    "preset": "@checkdigit/jest-config"
+  },
+  "eslintConfig": {
+    "extends": [
+      "@checkdigit/eslint-config"
+    ]
+  },
+  "files": [
+    "dist",
+    "SECURITY.md"
+  ]
 }
