+++ src/index.ts
@@ -1,10 +1,11 @@
 // index.ts
 
 /*
- * Copyright (c) 2021-2022 Check Digit, LLC
+ * Copyright (c) 2021-2023 Check Digit, LLC
  *
  * This code is licensed under the MIT license (see LICENSE.txt for details).
  */
 
 export { default } from './timeout';
-export * from './timeout';
+export { TimeoutError } from './error';
+export type { TimeoutOptions } from './options';
