+++ src/options.ts
@@ -1,0 +1,11 @@
+// options.ts
+
+/*
+ * Copyright (c) 2021-2023 Check Digit, LLC
+ *
+ * This code is licensed under the MIT license (see LICENSE.txt for details).
+ */
+
+export interface TimeoutOptions {
+  timeout?: number;
+}
