+++ tsconfig.json
@@ -4,9 +4,5 @@
     "rootDir": "src",
     "outDir": "build"
   },
-  "exclude": [
-    "node_modules",
-    "build",
-    "dist"
-  ]
+  "exclude": ["node_modules", "build", "coverage"]
 }
